+++ src/survey_ssr.h
@@ -175,14 +175,14 @@ void survey_ssr(const int N_day_screen_, const int N_aliquot_screen_,
       }
 
       {
-        const std::array<int, total_tp> out_n = count_summarise.get_total_obs();
+        const std::array<double, total_tp> out_n = count_summarise.get_total_obs();
         *(n_screen+outoffset) = out_n[screen_tp];
         *(n_pre+outoffset) = out_n[pre_tp];
         *(n_post+outoffset) = out_n[post_tp];
       }
 
       {
-        const std::array<int, total_tp> out_n_pos = count_summarise.get_total_pos();
+        const std::array<double, total_tp> out_n_pos = count_summarise.get_total_pos();
         *(n_pos_screen+outoffset) = out_n_pos[screen_tp];
         *(n_pos_pre+outoffset) = out_n_pos[pre_tp];
         *(n_pos_post+outoffset) = out_n_pos[post_tp];
